크롤링한 데이터가 이미 DB에 존재할 경우 업데이트하고, 크롤러를 개선한다.

+++ contract/contract-api/src/main/kotlin/product/ProductAddApi.kt
@@ -40,6 +40,6 @@ interface ProductAddApi {
 
     data class Response(
         @Schema(description = "등록된 상품의 개수", example = "10")
-        val savedItemsCount: Int
+        val savedItemsCount: Long
     )
 }
